backend/requirements.txt
Flask
flask-cors
SQLAlchemy
pydantic
python-dotenv
google-generativeai
